@@ -1,6 +1,6 @@
 {
   "name": "new-site",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "scripts": {
     "ng": "ng",
     "host": "node -p \"require('ip').address()\"",
@@ -13,27 +13,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "^7.2.0",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@angular/animations": "^8.0.2",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "^8.0.2",
+    "@angular/compiler": "~8.0.2",
+    "@angular/core": "^8.0.2",
+    "@angular/forms": "^8.0.2",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "~8.0.2",
+    "@angular/platform-browser-dynamic": "~8.0.2",
+    "@angular/router": "~8.0.2",
     "core-js": "^2.5.4",
     "hammerjs": "^2.0.8",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.5.2",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.0",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
+    "@angular-devkit/build-angular": "~0.800.3",
+    "@angular/cli": "~8.0.3",
+    "@angular/compiler-cli": "~8.0.2",
+    "@angular/language-service": "~8.0.2",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -48,6 +48,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "typescript": "~3.4.5"
   }
 }
